requirements.txt
numpy
requests
jupyter
jupyterlab
pandas
mysql-connector-python
pymongo
dnspython
ucimlrepo
scikit-learn
shap
matplotlib
openpyxl
seaborn
yfinance
pyodbc
flask
pydantic
fastapi
nltk
tensorflow
pandas_ta
statsmodels